+++ src/components/listPage/catDrop/CateDrop.tsx
@@ -1,15 +1,16 @@
 import * as t from "./cateDrop.style";
-import { useState } from "react";
 import { theme } from "../../../style/theme";
-import Select, { SelectChangeEvent } from "@mui/material/Select";
+import { useState } from "react";
 import MenuItem from "@mui/material/MenuItem";
 import FormControl from "@mui/material/FormControl";
+import Select, { SelectChangeEvent } from "@mui/material/Select";
 import { useAppDispatch, useAppSelector } from "../../../redux/store";
 import { editPagNo, editSort } from "../../../redux/reducer/listSlice";
 
 const CateDrop = () => {
   const [open, setOpen] = useState(false);
-  //-- reducer
+
+  //-- redux
   const { sort } = useAppSelector((state) => state.listSlice);
   const dispatch = useAppDispatch();
