+++ src/components/paymentPage/newPayInput/NewPayInput.tsx
@@ -1,11 +1,15 @@
 import * as t from "./newPayInput.style";
 import { useCallback, useState } from "react";
 import { useDaumPostcodePopup } from "react-daum-postcode";
-import { MainButton } from "../../../elements/buttons/Buttons";
-import Input from "../../../elements/input/Input";
 import { PayMethodSelect } from "../payMethod/PayMethod";
-import { useAppDispatch, useAppSelector } from "../../../redux/store";
+import Input from "../../../elements/input/Input";
+import { MainButton } from "../../../elements/buttons/Buttons";
 import { RootState } from "../../../redux/store";
+import { useAppDispatch, useAppSelector } from "../../../redux/store";
+import {
+  editIsName,
+  editisPNumber,
+} from "../../../redux/reducer/payCheckSlice";
 import {
   editDNo,
   editName,
@@ -14,10 +18,6 @@ import {
   editAddress1,
   editAddress2,
 } from "../../../redux/reducer/payUserSlice";
-import {
-  editIsName,
-  editisPNumber,
-} from "../../../redux/reducer/payCheckSlice";
 
 export const NewPayInput = () => {
   const { name, pNumber, zipcode, address1, address2 } = useAppSelector(
