+++ src/components/paymentPage/payRadioBtn/payRadioBtn.type.ts
@@ -1,0 +1,10 @@
+export interface RadiobtnType {
+  value?: number | string;
+  checked?: boolean;
+  type?: string;
+  select?: boolean;
+  top?: string;
+  bgColor?: string;
+  number?: number;
+  onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void;
+}
