+++ src/components/paymentPage/payTitle/PayTitle.tsx
@@ -1,4 +1,5 @@
 import * as t from "./payTitle.style";
+
 const PayTitle = () => {
   return <t.H1tag>결제하기</t.H1tag>;
 };
