+++ src/containers/detailPage/details/Details.tsx
@@ -17,7 +17,6 @@ const Details = () => {
     ignoreQueryPrefix: true,
   });
   const queryKeyword = String(query.keyword);
-  console.log(productNo);
 
   useGetDetailQuery(
     productNo,
