+++ src/containers/listPage/cardList/CardList.tsx
@@ -1,11 +1,10 @@
 import * as t from "./cardList.style";
-//pages//
-import Card from "../../../components/card/Card";
-import CardBadge from "../../../components/listPage/cardBadge/CardBadge";
-import CardIcon from "../../../components/listPage/cardIcon/CardIcon";
-import CateDrop from "../../../components/listPage/catDrop/CateDrop";
 import { CardListType } from "./cardList.type";
 import { SProductsType } from "../../searchPage/searchBody/searchPage.type";
+import Card from "../../../components/card/Card";
+import CateDrop from "../../../components/listPage/catDrop/CateDrop";
+import CardIcon from "../../../components/listPage/cardIcon/CardIcon";
+import CardBadge from "../../../components/listPage/cardBadge/CardBadge";
 
 const ListPage = (props: CardListType) => {
   return (
